Official Client for Jazz

How to get a Jazz server

Currently, in March 2018, Jazz is in the middle of major refactoring. The client is released mainly for supporting Jazz development, evaluation purposes and keeping a stable package name for the future. Unless you are interested in Jazz core development, we recommend you to wait until version 0.3.1 before giving Jazz a serious evaluation.

Run a docker image from docker hub (hub.docker.com)

This is fastest way for evaluation purposes. Check dockerhub to find the name of the latest image.

docker run -p8888:8888 kaalam/jazz_neat:0.2.1.99

Compile from source

The most controlable way is to compile the server from source.

git clone https://github.com/kaalam/Jazz
cd Jazz
./config.sh
cd server
make jazz
./jazz start

You can point a browser at http://localhost:8888/// to check if the server is running.